gen-builder.sh: Add libcrypt1
So far, Arch Linux-based distributions required the libxcrypto-compat
package to run the AppImage because libcrypt.so.1 is provided from a
different directory compared to Ubuntu. [1] Its file list [2] shows
this package bundles libcrypt.so.1, which is provided by the libcrypt1
package on Ubuntu focal. [3]

However, even if the libcrypt1 package is already installed into
$APPDIR/lib/x86_64-linux-gnu/ by dpkg-deb(1), for some unknown reason
appimage-builder moves libcrypt.so.1 and libcrypt.so.1.1.0 to
$APPDIR/runtime/compat/lib/x86_64-linux-gnu/, thus causing the game to
fail to find the libraries.

In order to work around this, a symbolic link is introduced on the
"after_runtime" step [4], since apparently the files are moved during
the "runtime" stage.

[1]: https://archlinux.org/packages/core/x86_64/libxcrypt-compat/
[2]: https://archlinux.org/packages/core/x86_64/libxcrypt-compat/files/
[3]: https://packages.ubuntu.com/focal/libcrypt1
[4]: https://appimage-builder.readthedocs.io/en/latest/reference/recipe.html?highlight=after_runtime#section-scripts